Changed Extensibility Wording to be Positive
When I first read this paragraph, it made me worry that bad things had been happening and instead of being excited about the new change, I was scared some bad thing may have crept into the source code that my come and get me later. It tried to re-word it to be positive, but not sure I understand the feature so may have gotten it wrong.