Add scoped service for type extensions to docs
When using EF core, I noticed that if I extended the Query with a class:

```c#
[ExtendObjectType("Query")]
public class CollectionsResolver {
    private readonly CollectionService collectionService;

    public CollectionsResolver([Service] CollectionService collectionService) =>
        this.collectionService = collectionService;

    public async Task<List<Collection>> GetCollections() =>
        await this.collectionService.GetCollections().ToListAsync();
}
```

every request after the 1st one would fail because the DbContext would get disposed. The solution was to register the CollectionsResolver as a scoped service: `services.AddScoped<CollectionsResolver>();`. I couldn't find this documented anywhere so I thought I should add it to the docs so others don't run into the same problem.