Update JsonObjectTypeExtensions.cs
Root cause: When GetDateTimeOffset() is performed on null, we will get an error stating "The requested operation requires an element of type 'String', but the target element has type 'Null' ". This will happen when the value of DateTime field is null. 

Code fix: The proposed change will ensure that the system returns null when DateTime field is null instead of throwing an error.